Fix: Refactor codes for ScriptItem.js file To improve the structure and apply Styled-components appropriately Ref: #134
